Add format validation for invoice attributes
PayEx have validations for the invoice attributes, we reflect those
validations here to avoid sending invalidig attributes to PayEx. This
patch also introduces a resolve option, where you can pass a Proc to try
to format the attribute correctly and retry the validation. For example
shorten the customer name if it's too long. See below for validations:

http://www.payexpim.com/technical-reference/pxorder/purchaseinvoiceprivate/